Freedom U540-C000 Bootloader Code
This commit contains the bootloader provided along with the Freedom
U540-C000 chip that is on SiFive's HiFive Unleashed board.  This has
been verified to boot on a HiFive Unleashed.  It contains both the
first-stage bootloader (which resides in SPI flash) and the zero-stage
bootloader (which resides in a mask ROM on the chip).

From this source it is possible to replace the entire contents of 0all
writable memories on the HiFive Unleashed with open source code.
Generating a ZSBL that exactly matches the contents of the ROM is an
execrise yet to be performed :).

Signed-off-by: Troy Benjegerdes <troy.benjegerdes@sifive.com>
Signed-off-by: Palmer Dabbelt <palmer@sifive.com>